feat:migrate to Vite, refactor architecture
