+++ src/components/AuthStatus.jsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from "react-router-dom";
-import { useAuth } from "../context/AuthProvider";
+import { useAuth } from "~/context/AuthProvider";
 
 export function AuthStatus() {
   const auth = useAuth();
