+++ src/components/Episode/EpisodeList.jsx
@@ -1,4 +1,4 @@
-import { useGetList } from "../../hooks/useGetList";
+import { useGetList } from "~/hooks/useGetList";
 
 export function EpisodeList() {
   const list = useGetList("https://rickandmortyapi.com/api/episode", "episodes");
