+++ src/components/Location/LocationList.jsx
@@ -1,4 +1,4 @@
-import { useGetList } from "../../hooks/useGetList";
+import { useGetList } from "~/hooks/useGetList";
 
 export function LocationList() {
   const list = useGetList("https://rickandmortyapi.com/api/location", "locations");
