+++ src/index.css
@@ -26,18 +26,8 @@ code {
   min-height: 100vh;
 }
 
-.home-page {
-  background: url('../public/home-background.jpg') top center no-repeat;
-  background-size: cover;
-  min-height: 100vh;
-}
-
 .nav__item {
   @apply font-bold text-rm-lime hover:opacity-80;
-}
-
-.character-img {
-  
 }
 
 .page-title {
@@ -45,15 +35,10 @@ code {
 }
 
 .episodes-page {
-  background: url('../public/episodes-background.png') top center no-repeat;
+  background: url('../episodes-background.png') top center no-repeat;
   background-size: cover;
   min-height: 100vh;
   padding: 20px 0;
-}
-
-.notfound-page {
-  background-color: #222222;
-  min-height: 100vh;
 }
 
 .episode-item {
@@ -74,12 +59,4 @@ code {
 .mobile-menu {
   @apply absolute top-0 right-0 left-0 bottom-0 h-screen px-5 pt-[70px] block sm:hidden;
   background-color: #222222;
-}
-
-
-.login-page {
-  background: url('../public/login.jpg') top center no-repeat;
-  background-size: cover;
-  min-height: 100vh;
-  padding: 20px 0;
 }
