--- src/pages/character/list/index.js
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const CharactersList = lazy(() => import('./CharactersList'));
