+++ src/pages/characters/detail/Character.jsx
@@ -1,0 +1,12 @@
+import { ErrorBoundary } from "~/components/ErrorBoundary/ErrorBoundary";
+import { CharacterItem } from "~/components/Character/CharacterItem";
+
+function Character() {
+  return (
+    <ErrorBoundary>
+      <CharacterItem />
+    </ErrorBoundary>
+  )
+}
+
+export default Character;
