+++ src/pages/characters/detail/index.js
