+++ src/pages/characters/list/PageCharactersList.jsx
@@ -1,0 +1,12 @@
+import { ErrorBoundary } from "~/components/ErrorBoundary/ErrorBoundary";
+import { CharacterList } from "~/components/Character/CharacterList";
+
+function PageCharactersList() { 
+  return (
+    <ErrorBoundary>
+      <CharacterList />
+    </ErrorBoundary>
+  );
+}
+
+export default PageCharactersList;
