+++ src/pages/characters/list/index.js
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export const PageCharactersList = lazy(() => import('./PageCharactersList'));
