--- src/pages/episode/list/index.js
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const EpisodesList = lazy(() => import('./EpisodesList'));
