+++ src/pages/episodes/list/index.js
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export const PageEpisodesList = lazy(() => import('./PageEpisodesList'));
