+++ src/pages/episodes/list/style.css
@@ -1,0 +1,6 @@
+.episodes-page {
+  background: url('../episodes-background.png') top center no-repeat;
+  background-size: cover;
+  min-height: 100vh;
+  padding: 20px 0;
+}
