+++ src/pages/home/Home.jsx
@@ -1,3 +1,5 @@
+import './style.css';
+
 export default function Home() {
   return (
     <div className="home-page">
