+++ src/pages/home/style.css
@@ -1,0 +1,5 @@
+.home-page {
+  background: url('../home-background.jpg') top center no-repeat;
+  background-size: cover;
+  min-height: 100vh;
+}
