--- src/pages/location/detail/Location.jsx
@@ -1,13 +1,0 @@
-import {ErrorBoundary} from "../../../components/ErrorBoundary/ErrorBoundary";
-import { LocationItem } from "../../../components/Location/LocationItem";
-
-function Location() {
-  return (
-    <ErrorBoundary>
-      <LocationItem />
-    </ErrorBoundary>
-  )
-}
-
-export default Location;
-
