--- src/pages/location/list/index.js
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const LocationsList = lazy(() => import('./LocationsList'));
