+++ src/pages/locations/detail/Location.jsx
@@ -1,0 +1,13 @@
+import {ErrorBoundary} from "~/components/ErrorBoundary/ErrorBoundary";
+import { LocationItem } from "~/components/Location/LocationItem";
+
+function Location() {
+  return (
+    <ErrorBoundary>
+      <LocationItem />
+    </ErrorBoundary>
+  )
+}
+
+export default Location;
+
