+++ src/pages/locations/detail/index.js
