+++ src/pages/locations/list/index.js
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export const PageLocationsList = lazy(() => import('./PageLocationsList'));
