+++ src/pages/login/style.css
@@ -1,0 +1,6 @@
+.login-page {
+  background: url('../login.jpg') top center no-repeat;
+  background-size: cover;
+  min-height: 100vh;
+  padding: 20px 0;
+}
