+++ src/pages/notfound/NotFound.jsx
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+import './style.css';
 
 function NotFound() {
