+++ src/pages/notfound/style.css
@@ -1,0 +1,4 @@
+.notfound-page {
+  background-color: #222222;
+  min-height: 100vh;
+}
