+++ vite.config.js
@@ -1,0 +1,23 @@
+import path from 'path';
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import tailwindcss from "tailwindcss";
+
+export default defineConfig(() => {
+  return {
+    resolve: {
+      alias: {
+        '~': path.resolve(__dirname, './src'),
+      },
+    },
+    build: {
+      outDir: 'build',
+    },
+    plugins: [react()],
+    css: {
+      postcss: {
+        plugins: [tailwindcss()],
+      },
+    },
+  };
+});
